Conflicts:
	integration/lifecycle/lifecycle_test.go
	integration/lifecycle/rlimit_test.go
	old/cgroups_manager/container_cgroups_manager.go
	old/linux_backend/bin/overlay.sh
	old/linux_backend/skeleton/stop.sh

+++ Godeps/_workspace/src/github.com/tedsuo/ifrit/ginkgomon/helpers.go
@@ -21,12 +21,16 @@ func Invoke(runner ifrit.Runner) ifrit.Process {
 	return process
 }
 
-func Interrupt(process ifrit.Process) {
-	process.Signal(os.Interrupt)
-	Eventually(process.Wait()).Should(Receive(), "interrupted ginkgomon process failed to exit in time")
+func Interrupt(process ifrit.Process, intervals ...interface{}) {
+	if process != nil {
+		process.Signal(os.Interrupt)
+		Eventually(process.Wait(), intervals...).Should(Receive(), "interrupted ginkgomon process failed to exit in time")
+	}
 }
 
-func Kill(process ifrit.Process) {
-	process.Signal(os.Kill)
-	Eventually(process.Wait()).Should(Receive(), "killed ginkgomon process failed to exit in time")
+func Kill(process ifrit.Process, intervals ...interface{}) {
+	if process != nil {
+		process.Signal(os.Kill)
+		Eventually(process.Wait(), intervals...).Should(Receive(), "killed ginkgomon process failed to exit in time")
+	}
 }
